requirements.txt
flask==3.0.3
flask_sqlalchemy==3.1.1
python-dotenv==1.0.1
gunicorn==22.0.0
openai==1.35.10
